Refactor banner configuration: consolidate and enhance validation logic for improved clarity and maintainability.
